requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
mysql-connector-python
google-generativeai
PyMySQL==1.1.0
cryptography==41.0.4
python-dotenv==1.0.0
reportlab==4.0.4
Flask-SQLAlchemy
google-generativeai
